Compiles and Runs with version 1.6.4

+++ org/pepsoft/worldpainter/WorldPainterView.java
@@ -20,7 +20,7 @@ public abstract class WorldPainterView extends JComponent {
 
     public abstract Point imageToWorldCoordinates(Point imageCoordinates);
 
-    public abstract Point viewToImageCoordinates(float x, float y);
+    public abstract Point viewToImageCoordinates(int x, int y);
 
     public abstract Point worldToImageCoordinates(int x, int y);
